requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
PyJWT==2.10.1
sqlparse==0.5.3
tomli==2.2.1
typing==3.7.4.3
typing_extensions==4.15.0
